requirements.txt
python-telegram-bot==20.7
PyGithub==2.1.1
langchain==0.1.0
langchain-community==0.0.13
huggingface-hub==0.20.1
qdrant-client==1.7.0
fastapi==0.109.0
uvicorn==0.25.0
python-dotenv==1.0.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0
